Attempt fix for logger consoleLog not completing before exit
Attaching Logger.consoleLog to process.on("exit") would not write to the console before the program exits. You are supposed to be able to execute sync operations of arbitrary length, including logging to the console. Even though everything up to and including the console.log statement was synchronous, perhaps it being inside a promise is what caused the issue.