Fix export not using updated class/interface names
I had renamed lots of classes and interfaces as part of the 2.0 release. Renaming the API constitutes a breaking change, hence the major version upgrade. However what I did not realise is that VSCode took care of aliasing everything in my index.ts to the old names, so actually the API didn't change at all. Great.